backend/requirements.txt
dj-database-url==0.2.1
Django==2.1.3
django-admin==1.3.2
django-cors-headers==2.4.0
django-excel-response2==2.0.8
django-heroku==0.3.1
django-six==1.0.4
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.6.1
pytz==2018.7
screen==1.0.1
whitenoise==4.1.2
xlwt==1.3.0
